added python code
This model consists of yahoo finance library which consists of the day-to-day price of the BITCOIN. It assists many works which actually include in the data preprocessing step. The prediction of the Bitcoin is fetched with the ARIMA-Auto Regressive Integrated Moving Average. This is too feasible to comprehend the system and invest into the stocks.